Made better use of existing components
