--- packages/widgets/src/components/Essay/index.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from "react"
-import StageSelector from "./StageSelector"
-import StageVisualizer from "./StageVisualizer"
-
-export default props => {
-  return (
-    <div>
-      <StageVisualizer {...props} />
-      <StageSelector {...props} />
-    </div>
-  )
-}
