+++ packages/widgets/src/components/Essays/EssayStageContainer.jsx
