+++ packages/widgets/src/components/Essays/StageVisualizer.jsx
